+++ src/components/ds-prototypes/Card/Card.js
@@ -8,32 +8,12 @@ import Link from 'next/link';
 import Image from 'next/image';
 
 function Card(props) {
-  const { name, imageUrl, description, url } = props;
+  const { children } = props;
 
   return (
-    <Fragment>
-      {imageUrl &&
-        <Image
-          alt=""
-          src={imageUrl}
-          layout="responsive"
-          width={900}
-          height={450}
-          quality={90}
-        />
-      }
-      {url &&
-        <h3 className={'heading'}>
-          <Link href={url}>
-            <a>{name}</a>
-          </Link>
-        </h3>
-      }
-      <div dangerouslySetInnerHTML={{
-          __html: description 
-        }}>
-      </div>
-    </Fragment>
+    <div className='ds-prototype-card'>
+      {children}
+    </div>
   );
 };
